src/agent_logic/core_functions.py
"""
Key Objectives:
1. Integrate new functionalities for executing Git operations (e.g., cloning repositories and comparing commits).
2. Refactor existing functions to accommodate the new script dependencies and maintain code coherence.
3. Provide utility functions for managing the output of git operations and differences.
4. Enhance error handling throughout the script to ensure robust performance.
5. Update inline documentation and comments to reflect any changes in logic or new methods introduced.
"""

from .git_operations import GitOperations
from .issue_analyzer import IssueAnalyzer
from .opord_analysis import OpordAnalysis

class AgentCore:
    def __init__(self):
        self.git_operations = GitOperations()
        self.issue_analyzer = IssueAnalyzer()
        self.opord_analysis = OpordAnalysis()

    def run(self):
        print("Agent is running")
        # Example of integrating new functionalities
        self.git_operations.clone_repository("https://github.com/example/repo.git")
        issues = self.issue_analyzer.fetch_open_issues("example/repo")
        self.opord_analysis.analyze_opord("docs/OPORD/OPORD-PR-CYBR-DOC-10.md")