tests/test_core_functions.py
"""
Key Objectives:
1. Add new test cases to cover functionalities introduced in the `git_operations.py` and `issue_analyzer.py`.
2. Ensure that existing tests are updated to reflect any changes in the methods of `core_functions.py`.
3. Implement mock objects where necessary to simulate git operations and issue fetching during testing.
4. Establish a clear coverage report to identify untested areas of the codebase.
5. Enhance error handling in tests to ensure they provide meaningful error messages in case of failure.
"""

import unittest
from unittest.mock import patch, MagicMock
from agent_logic.core_functions import AgentCore

class TestAgentCore(unittest.TestCase):
    def setUp(self):
        self.agent = AgentCore()

    @patch('agent_logic.git_operations.GitOperations.clone_repository')
    def test_git_clone_repository(self, mock_clone):
        # Simulate successful git clone
        mock_clone.return_value = None
        self.agent.git_operations.repo_url = "https://github.com/example/repo.git"
        self.agent.git_operations.clone_repository()
        mock_clone.assert_called_once()

    @patch('agent_logic.git_operations.GitOperations.fetch_commit_diff')
    def test_git_fetch_commit_diff(self, mock_fetch_diff):
        # Simulate fetching commit diff
        mock_fetch_diff.return_value = "diff --git a/file.txt b/file.txt"
        diff = self.agent.git_operations.fetch_commit_diff("commit1", "commit2")
        mock_fetch_diff.assert_called_once_with("commit1", "commit2")
        self.assertIn("diff --git", diff)

    @patch('agent_logic.issue_analyzer.IssueAnalyzer.fetch_open_issues')
    def test_issue_fetch_open_issues(self, mock_fetch_issues):
        # Simulate fetching open issues
        mock_fetch_issues.return_value = [{"title": "Issue 1"}, {"title": "Issue 2"}]
        issues = self.agent.issue_analyzer.fetch_open_issues()
        mock_fetch_issues.assert_called_once()
        self.assertEqual(len(issues), 2)

    @patch('agent_logic.issue_analyzer.IssueAnalyzer.analyze_issues')
    def test_issue_analyze_issues(self, mock_analyze_issues):
        # Simulate analyzing issues
        mock_analyze_issues.return_value = {"total_issues": 2, "labels": {}, "assignees": {}}
        analysis = self.agent.issue_analyzer.analyze_issues([{"title": "Issue 1"}, {"title": "Issue 2"}])
        mock_analyze_issues.assert_called_once()
        self.assertEqual(analysis["total_issues"], 2)

    @patch('agent_logic.opord_analysis.OpordAnalysis.parse_opord')
    def test_opord_parse_opord(self, mock_parse_opord):
        # Simulate parsing OPORD
        mock_parse_opord.return_value = "## Task Organization:\n- Task 1\n##"
        content = self.agent.opord_analysis.parse_opord()
        mock_parse_opord.assert_called_once()
        self.assertIn("Task 1", content)

if __name__ == '__main__':
    unittest.main()